feat(front): adds my-projects page (#81)
## Ticket

**Title: ETQM, j’ai une page “mes projets” pour pouvoir consulter mes
projets de codes enregistrés**
**Notion Link:
https://www.notion.so/jkergal/ETQM-j-ai-une-page-mes-projets-pour-pouvoir-consulter-mes-projets-de-codes-enregistr-s-6fe55c17a2ec4cdf88f367431d296fdb?pvs=4**

## Description
- querying data
- displaying data
- adds filter bar
- adds search bar
- reponsive


![image](https://github.com/reynan-dev/codenaute/assets/89657288/47d0996d-21c3-490a-a868-4bf12727e843)

![image](https://github.com/reynan-dev/codenaute/assets/89657288/920c18aa-84c7-4598-901f-e03524fa43be)

![image](https://github.com/reynan-dev/codenaute/assets/89657288/631aa5dc-c966-4199-bc7c-5659715cffd2)

![image](https://github.com/reynan-dev/codenaute/assets/89657288/9e3ceaaf-4630-415d-985e-de8b51f1d238)

![image](https://github.com/reynan-dev/codenaute/assets/89657288/18e09366-4a13-48a8-b7b1-10c3645dcbf8)


## Guidelines

🔗 Link your PRs to their related Notion tickets
✍️ Use the commit convention to your pull request title
📃 If your code affects the build, update `README.md`
🚫 Do not merge a PR until all comments are resolved
🗑 Remove your branch after merging

+++ client/src/pages/code/_components/custom-file-explorer.tsx
@@ -9,6 +9,7 @@ import { renameFolder } from 'pages/code/_helpers/rename-folder';
 import { Position, useContextMenuEvents } from 'pages/code/_hooks/use-context-menu-events';
 import React, { useContext, useEffect, useRef, useState } from 'react';
 import { AiFillFolder, AiFillFolderOpen, AiOutlineFile } from 'react-icons/ai';
+import { toast } from 'react-toastify';
 import { twJoin, twMerge } from 'tailwind-merge';
 
 interface CustomFileExplorerProps {
@@ -26,8 +27,13 @@ export const CustomFileExplorer = ({ className, files }: CustomFileExplorerProps
 	const [newFolderName, setNewFolderName] = useState<string | null>(null);
 	const { sandpack } = useSandpack();
 	const filesElementRef = useRef<HTMLDivElement | null>(null);
-	const { setActiveFile, setVisibleFiles, visibleFiles, setCurrentProjectData } =
-		useContext(ProjectContext);
+	const {
+		setActiveFile,
+		setVisibleFiles,
+		visibleFiles,
+		setCurrentProjectData,
+		currentProjectData
+	} = useContext(ProjectContext);
 
 	useContextMenuEvents(filesElementRef, setContextMenuPosition);
 
@@ -63,6 +69,10 @@ export const CustomFileExplorer = ({ className, files }: CustomFileExplorerProps
 				handleRenameStart(selectedNode, event);
 			}
 			if (action === 'delete' && selectedNode !== null) {
+				if (selectedNode.path === currentProjectData?.main)
+					return toast.error('You cannot delete the main file of your project', {
+						autoClose: 10000
+					});
 				sandpack.deleteFile(selectedNode.path);
 			}
 		};
